Views

+++ app/model/User.php
@@ -2,6 +2,7 @@
 namespace App\model;
 
 use App\model\Model;
+use Exception;
 /**
  * Class Model responsavel pelas requisições da tavela usuario
  * @version ${2:2.0.0
@@ -14,9 +15,15 @@ class User extends Model{
      * @return bool
      */
     public function create(array $param) {
-       $insert = $this->conect->prepare("INSERT INTO usuario(name, email, password) VALUES(:name, :email, :password)");
-       if($insert->execute($param)) return true;
-        return false;
+        //verifica  se não algum erro na conexão.
+        if(gettype($this->conect) == "object") {
+            //perarando o sql a ser executado
+            $insert = $this->conect->prepare("INSERT INTO usuario(name, email, password) VALUES(:name, :email, :password)");
+            //executa o sql e verifica se deu aldo de errado
+            if($insert->execute($param)) return true;
+            throw new Exception("[ATENÇÃO]Erro de execução", 30);
+        }
+        return $this->conect;//retorna o erro caso haja.
     }
     /**
      * Método para atualizar usuario
@@ -25,9 +32,15 @@ class User extends Model{
      * @return bool
      */
     public function update(array $param) {
-        $insert = $this->conect->prepare("UPDATE usuario SET name= :name, email= :email, password= :password WHERE iduser= :id");
-        if($insert->execute($param)) return true;
-         return false;
+         //verifica  se não algum erro na conexão.
+        if(gettype($this->conect) == "object") {
+            //perarando o sql a ser executado
+            $insert = $this->conect->prepare("UPDATE usuario SET name= :name, email= :email, password= :password WHERE iduser= :id");
+            //executa o sql e verifica se deu aldo de errado
+            if($insert->execute($param)) return true;
+            throw new Exception("[ATENÇÃO]Erro de execução", 30);
+        }
+        return $this->conect;//retorna o erro caso haja.
     }
     /**
      * Método para deletar usuario
@@ -36,18 +49,30 @@ class User extends Model{
      * @return bool
      */
     public function delete(array $param) {
-        $insert = $this->conect->prepare("DELETE FROM usuario WHERE iduser= :id");
-        if($insert->execute($param)) return true;
-        return false;
+        //verifica  se não algum erro na conexão.
+        if(gettype($this->conect) == "object") {
+            //perarando o sql a ser executado
+            $insert = $this->conect->prepare("DELETE FROM usuario WHERE iduser= :id");
+            //executa o sql e verifica se deu aldo de errado
+            if($insert->execute($param)) return true;
+            throw new Exception("[ATENÇÃO]Erro de execução", 30);
+        }
+        return $this->conect;//retorna o erro caso haja.
     }
     /**
      * Método para realizar select de todos os usuario
      *
-     * @return bool
+     * @return mixed
      */
     public function getAll() {
-        $get = $this->conect->query("SELECT * FROM usuario");
-        if($result = $get->fetch()) return $result;
-        return false;
+        //verifica  se não algum erro na conexão.
+        if(gettype($this->conect) == "object") {
+            //Execução da query sql
+            $get = $this->conect->query("SELECT * FROM usuario");
+            //verifica e trata o resultado da query
+            if($result = $get->fetch()) return $result;//retorna o resultado da query
+            throw new Exception("[ATENÇÃO]Erro de execução", 30);//retorna falso caso haja algum erro
+        }
+        return $this->conect;//retorna o erro caso haja.
     }
 }
